more docs fixes and move parse_log_files function
